improve script

+++ scripts/bump_ddtrace.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import platform
 import re
 import subprocess
 import sys
@@ -9,6 +10,14 @@ import requests
 PROJECT = ""
 GH_USERNAME = ""
 OS_USERNAME = ""
+
+
+PARTIAL_BRANCH_NAME = "bump-ddtrace-to-"
+PACKAGE_NAME = "ddtrace"
+
+linux = platform.system() == "Linux"
+HOME_PREFIX = "/home" if linux else "/Users"
+OPEN_CMD = "xdg-open" if linux else "open"
 
 
 def run_command(command, check=True, cwd=None):
@@ -43,40 +52,48 @@ def get_latest_dependency_version(package_name):
     return None
 
 
-def modify_requirements_file(requirements_path, package_name, new_version):
+def modify_requirements_file(requirements_path, new_version):
     """Step 2: Modify the requirements.in file."""
-    print(f"Updating {package_name} to version {new_version} in {requirements_path}")
+    print(f"Updating {PACKAGE_NAME} to version {new_version} in {requirements_path}")
     with open(requirements_path, "r") as file:
         lines = file.readlines()
 
     with open(requirements_path, "w") as file:
         for line in lines:
-            if re.match(rf"{package_name}==[0-9]+(\.[0-9]+)*", line):
-                file.write(f"{package_name}=={new_version}\n")
+            if re.match(rf"{PACKAGE_NAME}==[0-9]+(\.[0-9]+)*", line):
+                file.write(f"{PACKAGE_NAME}=={new_version}\n")
             else:
                 file.write(line)
 
 
-def run_bazel_commands(project_path):
+def run_bazel_commands(project_path, only_vendor=False):
     """Step 3: Run necessary Bazel commands."""
     commands = ["bzl run //:requirements.update", "bzl run //:requirements_vendor"]
+    if only_vendor:
+        commands = commands[1:]
     for cmd in commands:
         run_command(cmd, cwd=project_path)
 
 
-def commit_and_push_changes(branch_name, package_name, project_path):
+def commit_and_push_changes(branch_name, project_path, latest_version):
     """Step 4: Commit the changes and push to remote."""
     run_command("git add requirements*", cwd=project_path)
-    run_command(f"git commit -m 'Bump {package_name} version'", cwd=project_path)
+    input("Are you ready to commit and push the changes? Press Enter to continue...")
+    run_command(f"git commit -m 'Bump {PACKAGE_NAME} version to {latest_version}'", cwd=project_path)
     run_command(f"git push --set-upstream origin {branch_name}", cwd=project_path)
 
 
-def create_bump_pr(branch_name, package_name, project_path, requirements_path, latest_version, project):
-    create_git_branch(branch_name, project_path)
-    modify_requirements_file("/".join((project_path, requirements_path)), package_name, latest_version)
-    run_bazel_commands(project_path)
-    commit_and_push_changes(branch_name, package_name, project_path)
-    run_command(f"open https://github.com/DataDog/{project}/pull/new/{branch_name}")
+def create_bump_pr(branch_name, project_path, requirements_path, latest_version, project):
+    try:
+        create_git_branch(branch_name, project_path)
+        modify_requirements_file("/".join((project_path, requirements_path)), latest_version)
+        run_bazel_commands(project_path)
+        commit_and_push_changes(branch_name, project_path, latest_version)
+        run_command(f"{OPEN_CMD} https://github.com/DataDog/{project}/pull/new/{branch_name}")
+    except Exception:
+        run_command("git reset --hard HEAD", cwd=project_path)
+        run_command("git checkout main", cwd=project_path)
+        run_command(f"git branch -D {branch_name}", cwd=project_path)
 
 
 def update_pr(branch_name, project_path):
@@ -88,46 +105,60 @@ def update_pr(branch_name, project_path):
     run_command(f"git checkout {branch_name}", cwd=project_path)
     # save git sha
     git_sha = run_command("git rev-parse HEAD", cwd=project_path).stdout.strip()
-    # reset branch to main_branch
-    run_command(f"git reset --hard {main_branch_name}", cwd=project_path)
-    # cherry-pick commit
-    run_command(f"git cherry-pick {git_sha}", cwd=project_path)
-    # reset requirement_* files to main state
-    run_command(f"git checkout {main_branch_name} -- requirements_*", cwd=project_path)
-    # run bazel commands
-    run_bazel_commands(project_path)
-    # add new changes
-    run_command("git add requirements_*", cwd=project_path)
-    # cherry-pick continue with changes, no editor
-    run_command("git cherry-pick --continue --no-edit", cwd=project_path)
-    # force push changes
-    run_command("git push --force", cwd=project_path)
+    cherry_pick_success = False
+    try:
+        # reset branch to main_branch
+        run_command(f"git reset --hard {main_branch_name}", cwd=project_path)
+        # cherry-pick commit, this will fail if there are conflicts, that's OK
+        result = run_command(f"git cherry-pick {git_sha}", check=False, cwd=project_path)
+        if "conflict" in result.stdout.lower() or "conflict" in result.stderr.lower():
+            # reset requirement_* files to main state
+            run_command(f"git checkout {main_branch_name} -- requirements_*", cwd=project_path)
+            # run bazel commands
+            run_bazel_commands(project_path, only_vendor=True)
+            # add new changes
+            run_command("git add requirements_*", cwd=project_path)
+            # cherry-pick continue with changes, no editor
+            run_command("git -c core.editor=true cherry-pick --continue", check=False, cwd=project_path)
+            cherry_pick_success = True
+        else:
+            cherry_pick_success = True
+
+    except Exception:
+        # if there are conflicts, abort the cherry-pick and reset the branch
+        run_command("git cherry-pick --abort", check=False, cwd=project_path)
+        run_command(f"git reset --hard {git_sha}", cwd=project_path)
+
+    if cherry_pick_success:
+        # force push changes
+        run_command("git push --force", cwd=project_path)
 
 
-def main():
+def main(package_version=None):
     if not all((GH_USERNAME, OS_USERNAME, PROJECT)):
         print("Fill the required constants at the top of the file (project and usernames)")
         sys.exit(1)
 
-    partial_branch_name = "bump-ddtrace-to-"
-    package_name = "ddtrace"
     requirements_path = "requirements.in"
-    project_path = f"/Users/{OS_USERNAME}/go/src/github.com/DataDog/{PROJECT}"
+    project_path = f"{HOME_PREFIX}/{OS_USERNAME}/go/src/github.com/DataDog/{PROJECT}"
 
-    latest_version = get_latest_dependency_version(package_name)
-    branch_name = "/".join((GH_USERNAME, partial_branch_name + latest_version))
+    latest_version = package_version or get_latest_dependency_version(PACKAGE_NAME)
+    branch_name = "/".join((GH_USERNAME, PARTIAL_BRANCH_NAME + latest_version))
 
     # To decide whether if create or update the PR, check if the branch name exists in origin
-    branch_exists = (
-        run_command(f"git ls-remote --heads origin {branch_name}", check=False, cwd=project_path).returncode == 0
-    )
-    create_pr = not branch_exists
+    result = run_command(f"git ls-remote --heads origin {branch_name}", check=False, cwd=project_path)
+    branch_exists = result.stdout
 
-    if create_pr:
-        create_bump_pr(branch_name, package_name, project_path, requirements_path, latest_version, PROJECT)
+    if not branch_exists:
+        create_bump_pr(branch_name, project_path, requirements_path, latest_version, PROJECT)
     else:
         update_pr(branch_name, project_path)
 
 
 if __name__ == "__main__":
-    main()
+    # Usage: ./bump_ddtrace.py 0.1.2
+    # Get package version from command line arguments
+    package_version = None
+    if len(sys.argv) > 1:
+        package_version = sys.argv[1]
+    main(package_version)
